stream: move streamVideo functions to stream
